requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
maya==0.6.1
mdurl==0.1.2
multidict==6.0.4
pendulum==2.1.2
pydantic==1.10.8
PyJWT==2.7.0
PyMySQL==1.0.3
rich==13.3.5
Werkzeug==2.3.4
